fixes 31 OCT
Added a new logic in the recipe deletion function which deletes the related products of that recipe.

added the fields and logic for the fields of the product modal.

Learned about custom scroll bar libraries and chose to use overlayscrollbars library.

Implemented the custom scroll bar as a whole.

Tried to implement the custom scroll bar in the recipe list but still not working as intended ( mostly due to front end complications )